Close the plstore handle after use
Before this, the plstore stores a handle to `oauth2-token-file' and it
is kept open at all times.  This may cause issues when the file is
accessed by other processes and it may confuse Emacs. This patch
ensures that the plstore is properly closed upon each use.

Note that after this change, the plstore field in oauth2-token is
unused and will always be nil, but is kept for backward-compatibility
for now.

* packages/oauth2/oauth2.el (oauth2--with-plstore): New macro.
* packages/oauth2/oauth2.el (oauth2-refresh-token)
(oauth2-auth-and-store): Use it.