Implement OAuth2 PKCE extension (RFC7636)
Proof Key for Code Exchange is an extension to prevent CSRF and
authorization code injection attacks.  This is implemented in other
OAuth2 providers, e.g. thunderbird, mutt_oauth2.py, etc.  On testing
with predefined credentials, it looks like Outlook requires this
extension for requesting access-token or the connection will be denied
even with a retrieved access-token.  This is opt-in, and is enabled
when passing use-pkce as non-nil.

* packages/oauth2/oauth2.el (oauth2-token): Add code-verifier slot.
* packages/oauth2/oauth2.el (oauth2--generate-code-verifier,
oauth2--get-challenge-from-verifier): Add.
* packages/oauth2/oauth2.el (oauth2--update-plstore): Store
code-verifier in plstore.
* packages/oauth2/oauth2.el (oauth2-request-authorization,
oauth2-request-access, oauth2-auth): Add `code-verifier' parameter and
pass down or add to request URL data.
* packages/oauth2/oauth2.el (oauth2-auth-and-store): Generate
`code-verifier' and pass down.
* packages/oauth2/oauth2-tests.el: Add simple unit tests.