Fix cache handling
During testing, it is observed that while the `refresh-token' can be
reused after the initial request, the same `access-token' cannot be
shared between different hosts, e.g. IMAP and SMTP servers.  If one
gets the `access-token' when receiving emails from an IMAP server and
reuse the same `access-token' to send mail to an SMTP server the
login will fail.  The same happens the other way around.

As a result, the access tokens for different servers need to be stored
separately, and each should store its own request timestamp for
expiration checking.

It is also possible that the credentials in the cache are invalid due
to errors.  In such case it should request a new token anyway.

For backward compatibility, the `host-name' parameter is optional, and
if it is not provided, the cache is effectively disabled and each
request will refresh the `access-token' to ensure it works.  Also,
`oauth2-token-access-token' will be updated on each token refresh to
be the latest token retrieved regardless of the host.

* packages/oauth2/oauth2.el (oauth2--update-request-cache,
oauth2--get-from-request-cache): Add helper functions; define the
request-cache structure.
* packages/oauth2/oauth2.el (oauth2-token): Replace request-timestamp
with request-cache structure.
* packages/oauth2/oauth2.el (oauth2--update-plstore): Store the new
request-cache structure.
* packages/oauth2/oauth2.el (oauth2-auth-and-store): Add host-name
parameter and pass down.  Also validate cache or request new.